requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.4.0
faiss-cpu>=1.11.0
rank-bm25==0.2.2
openai==1.45.0
httpx==0.27.0
pypdf==4.3.1
trafilatura==1.12.2
beautifulsoup4==4.12.3
rich==13.7.1
streamlit==1.37.1
